requirements.txt
Alfred-Workflow==1.11.1
socketIO-client==0.6.1
